Relicense all .yml files as CC-1.0
The primary license of the project is changing to:
  AGPL-3.0-or-later WITH Web-Template-Output-Additional-Permission-3.0-or-later

with some specific files to be licensed under the one of two licenses:
   CC0-1.0
   LGPL-3.0-or-later

This commit is one of the many steps to relicense the entire codebase.

We decided to relicense the .yml files, which currently contain only
translations, under CC0-1.0 for now.  The project may in future place these
files under a copyleft license, but will discuss that with the contributor
community before doing so.

Documentation granting permission for this relicensing (from all past
contributors who hold copyrights) is on file with Software Freedom
Conservancy, Inc.