Relicense all .css files as LGPL-3.0-or-later
The primary license of the project is changing to:
  AGPL-3.0-or-later WITH Web-Template-Output-Additional-Permission-3.0-or-later

The Additional Permission is designed to permit publicly distributed CSS to
be relicensed under LGPL-3.0-or-later, so we're also relicensing all CSS
files in the repository as LGPL-3.0-or-later, too.

This commit is one of the many steps to relicense the entire codebase.

Documentation granting permission for this relicensing (from all past
contributors who hold copyrights) is on file with Software Freedom
Conservancy, Inc.