Add validation to nonprofits state code on create and update.
The create nonprofit command wasn't validating the state code for US states only, when trying to acess the application the route match fails.

To solve the problem a validation was added to the nonprofit model, using the 'countries' gem to get the states.

A hardcoded states list could be used to validates the states or another gem, but the 'countries' gem was already used in the project and it's widely used in the community.

Co-authored-by: igor-paiva <igorbatistapaiva@outlook.com>
Co-authored-by: thiagolopess <thiago.lopes.santos.tls@gmail.com>