Pull out unneeded change to puma port config
At one point had a bind call in the puma config with the port on it. Created `tcp_port` so the env var didn't need to be read twice. But found out the bind that was being specified was essentially the same as default and removed it. This should have been cleaned up at the time but was missed.