Provide to run worker as normal outside Docker
When using Docker, it is useful to have the job logs separate and so it makes sense to run the job worker in a separate container. For a dev running the project on bare metal (without Docker), they should have an easy one-command process as it had been with bin/dev prior. This maintains that by making bin/dev switch which Procfile it uses based on an ENV var.