Switch back but make time zone select helper
Reverted the change to storing Rails timezone names in the database due to some of the custom query code directly working with timezones in postgres.

Opted instead for a helper that outputs the timezones in a similar fashion for the end user in the select even though it is not the Rails helper, it can be used for when events get timezones added.